requirements.txt
django==1.11.8
django-crispy-forms>=1.6.0
django-import-export>=0.5.1
django-reversion>=2.0.0
django-simple-captcha==0.5.9
django-formtools==2.1
django-widget-tweaks==1.4.3
django-pure-pagination==0.3.0
future==0.15.2
httplib2==0.9.2
pytz
six==1.10.0
redis==2.10.6
psutil==5.4.7
mysqlclient
scrapy>=1.5.1
python-scrapyd-api==2.1.2
requests==2.19.1
html5lib==1.0.1
redis==2.10.6
pymysql==0.9.2
pymongo==3.7.1
pillow==5.2.0
paramiko==2.4.2
celery==4.2.1
kombu==4.2.1

